chore: restructure polkadot-node-jaeger
